refactor(messaging): rename transport/encoding package to rfc26
The package implements exactly one thing — the WakuMessage version=1
payload codec from RFC 26/WAKU2-PAYLOAD — so name it after the spec.
"encoding" was both vague and a near-homonym of the neighbouring
"encryption" (X3DH) layer; "rfc26" is unambiguous and self-documenting.

- pkg/.../transport/encoding -> pkg/.../transport/rfc26 (package rfc26),
  encoding.go -> rfc26.go, encoding_test.go -> rfc26_test.go.
- All importers and identifier references updated
  (encoding.X -> rfc26.X) across transport, waku/api, waku/common, and
  the doc references in waku/types and messaging_api.
- Added rfc26/README.md linking the spec (https://rfc.vac.dev/spec/26/),
  its EIP-627 / RLPx-ECIES origins, and the pm#420 removal tracking.

Pure rename; no behaviour or wire-format change.

Refs status-im/status-go#7462
