+++ pkg/messaging/waku/common/filter_test.go
@@ -17,7 +17,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
 
-	"github.com/status-im/status-go/pkg/messaging/layers/transport/encoding"
+	"github.com/status-im/status-go/pkg/messaging/layers/transport/rfc26"
 )
 
 const testShard = "/waku/2/rs/16/32"
@@ -187,12 +187,12 @@ func cloneFilter(orig *Filter) *Filter {
 }
 
 func generateCompatibleReceivedMessage(t *testing.T, f *Filter) *ReceivedMessage {
-	keyInfo := &encoding.KeyInfo{}
-	keyInfo.Kind = encoding.Symmetric
+	keyInfo := &rfc26.KeyInfo{}
+	keyInfo.Kind = rfc26.Symmetric
 	keyInfo.SymKey = f.KeySym
 
 	var version uint32 = 1
-	p := new(encoding.Payload)
+	p := new(rfc26.Payload)
 	p.Data = make([]byte, 20)
 	_, err := crand.Read(p.Data) // nolint: gosec
 	require.NoError(t, err)
